requirements.txt
ansicolors==1.1.8
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.2.1
comm==0.2.2
conda==22.11.1
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fair-esm==2.0.0
fastjsonschema==2.21.1
fonttools==4.58.2
idna @ file:///croot/idna_1666125576474/work
ihm==2.2
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
modelcif==1.2
msgpack==1.1.0
nbclient==0.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.2.6
omegaconf==2.3.0
packaging==25.0
pandas==2.3.0
papermill==2.6.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pip==22.3.1
platformdirs==4.3.8
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pygments==2.19.2
pyopenssl @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.2.3
pyright==1.1.402
pysocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytz==2025.2
pyyaml==6.0
pyzmq==27.0.0
referencing==0.36.2
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rpds-py==0.26.0
ruamel-yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel-yaml-clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
scipy==1.15.3
seaborn==0.13.2
setuptools==65.5.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.6.3
tenacity==9.1.2
toolz @ file:///croot/toolz_1667464077321/work
tornado==6.5.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.14.3
typing-extensions==4.14.0
urllib3 @ file:///croot/urllib3_1670526988650/work
wcwidth==0.2.13
wheel==0.37.1
